+++ src/index.ts
@@ -29,6 +29,7 @@ export type {
 export { useTextHeight } from './hooks/useTextHeight'
 export { usePreparedText } from './hooks/usePreparedText'
 export { useFlashListHeights } from './hooks/useFlashListHeights'
+export type { FlashListHeightsResult } from './hooks/useFlashListHeights'
 export { useStreamingLayout } from './hooks/useStreamingLayout'
 export { useMultiStreamLayout } from './hooks/useMultiStreamLayout'
 export { measureHeights, measureTokenWidth } from './prepare'